# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 2.0.0
  description: TransportPro - Transportation Business Management API
servers:
  - url: /api
...
paths:
...
  /drivers:
    get:
      operationId: getDrivers
      tags: [drivers]
      summary: List drivers
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Driver"
    post:
      operationId: createDriver
      tags: [drivers]
      summary: Create driver
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
  /drivers/{id}:
    get:
      operationId: getDriver
      tags: [drivers]
      summary: Get driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
    put:
      operationId: updateDriver
      tags: [drivers]
      summary: Update driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DriverInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Driver"
    delete:
      operationId: deleteDriver
      tags: [drivers]
      summary: Delete driver
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
...
  /trips/{id}/expenses:
    get:
      operationId: getTripExpenses
      tags: [expenses]
      summary: Get trip expenses
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TripExpense"
    post:
      operationId: createTripExpense
      tags: [expenses]
      summary: Add trip expense
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TripExpenseInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TripExpense"
...
components:
  schemas:
...
    Driver:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        passportNumber:
          type: string
          nullable: true
        licenseNumber:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, suspended, terminated]
        statusEffectiveDate:
          type: string
          nullable: true
        monthlySalary:
          type: number
        assignedTruckId:
          type: integer
          nullable: true
        assignedTruckPlate:
          type: string
          nullable: true
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, status, monthlySalary, createdAt]
    DriverInput:
      type: object
      properties:
        name:
          type: string
        passportNumber:
          type: string
          nullable: true
        licenseNumber:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        status:
          type: string
          enum: [active, suspended, terminated]
        statusEffectiveDate:
          type: string
          nullable: true
        monthlySalary:
          type: number
        assignedTruckId:
          type: integer
          nullable: true
        notes:
          type: string
          nullable: true
      required: [name, status, monthlySalary]
...
    TripExpense:
      type: object
      properties:
        id:
          type: integer
        tripId:
          type: integer
        costType:
          type: string
        description:
          type: string
          nullable: true
        amount:
          type: number
        currency:
          type: string
        createdAt:
          type: string
      required: [id, tripId, costType, amount, currency, createdAt]
    TripExpenseInput:
      type: object
      properties:
        costType:
          type: string
        description:
          type: string
          nullable: true
        amount:
          type: number
        currency:
          type: string
          default: USD
      required: [costType, amount]
...